v0.3.2: significantly refactored custom tile visuals
